se reacomodan métodos en base a la consigna en productosservice
