+++ socialMeli/src/main/java/com/javabootcamp/socialmeli/service/IProductService.java
@@ -1,9 +1,11 @@
 package com.javabootcamp.socialmeli.service;
 
+import com.javabootcamp.socialmeli.dto.LastPostDto;
 import com.javabootcamp.socialmeli.dto.ProductDto;
 import com.javabootcamp.socialmeli.model.Product;
 
 public interface IProductService {
     Product addProduct(Product product);
     ProductDto productEntityToDto(Product product);
+    LastPostDto getPostFromLastTwoWeeks(Integer userId);
 }
